refactor out transaction code
