+++ src/main/java/bio/terra/service/dataset/DatasetService.java
@@ -51,7 +51,7 @@ import bio.terra.service.resourcemanagement.azure.AzureStorageAccountResource;
 import bio.terra.service.snapshot.exception.AssetNotFoundException;
 import bio.terra.service.tabulardata.azure.StorageTableService;
 import bio.terra.service.tabulardata.google.bigquery.BigQueryDatasetPdao;
-import bio.terra.service.tabulardata.google.bigquery.BigQueryPdao;
+import bio.terra.service.tabulardata.google.bigquery.BigQueryTransactionPdao;
 import bio.terra.stairway.ShortUUID;
 import com.azure.storage.blob.sas.BlobSasPermission;
 import com.fasterxml.jackson.core.JsonProcessingException;
@@ -77,7 +77,7 @@ public class DatasetService {
   private final LoadService loadService;
   private final ProfileDao profileDao;
   private final StorageTableService storageTableService;
-  private final BigQueryPdao bigQueryPdao;
+  private final BigQueryTransactionPdao bigQueryTransactionPdao;
   private final BigQueryDatasetPdao bigQueryDatasetPdao;
   private final MetadataDataAccessUtils metadataDataAccessUtils;
   private final ResourceService resourceService;
@@ -93,7 +93,7 @@ public class DatasetService {
       LoadService loadService,
       ProfileDao profileDao,
       StorageTableService storageTableService,
-      BigQueryPdao bigQueryPdao,
+      BigQueryTransactionPdao bigQueryTransactionPdao,
       BigQueryDatasetPdao bigQueryDatasetPdao,
       MetadataDataAccessUtils metadataDataAccessUtils,
       ResourceService resourceService,
@@ -106,7 +106,7 @@ public class DatasetService {
     this.loadService = loadService;
     this.profileDao = profileDao;
     this.storageTableService = storageTableService;
-    this.bigQueryPdao = bigQueryPdao;
+    this.bigQueryTransactionPdao = bigQueryTransactionPdao;
     this.bigQueryDatasetPdao = bigQueryDatasetPdao;
     this.metadataDataAccessUtils = metadataDataAccessUtils;
     this.resourceService = resourceService;
@@ -394,7 +394,7 @@ public class DatasetService {
   public List<TransactionModel> enumerateTransactions(UUID id, long limit, long offset) {
     Dataset dataset = retrieve(id);
     try {
-      return bigQueryPdao.enumerateTransactions(dataset, limit, offset);
+      return bigQueryTransactionPdao.enumerateTransactions(dataset, limit, offset);
     } catch (InterruptedException e) {
       throw new RuntimeException("Error enumerating transactions", e);
     }
@@ -403,7 +403,7 @@ public class DatasetService {
   public TransactionModel retrieveTransaction(UUID id, UUID transactionId) {
     Dataset retrieve = retrieve(id);
     try {
-      return bigQueryPdao.retrieveTransaction(retrieve, transactionId);
+      return bigQueryTransactionPdao.retrieveTransaction(retrieve, transactionId);
     } catch (InterruptedException e) {
       throw new RuntimeException("Error retrieving transaction", e);
     }
