+++ src/main/java/bio/terra/service/dataset/flight/ingest/IngestInsertIntoDatasetTableStep.java
@@ -11,7 +11,7 @@ import bio.terra.service.dataset.DatasetTable;
 import bio.terra.service.dataset.flight.transactions.TransactionUtils;
 import bio.terra.service.job.JobMapKeys;
 import bio.terra.service.tabulardata.google.bigquery.BigQueryDatasetPdao;
-import bio.terra.service.tabulardata.google.bigquery.BigQueryPdao;
+import bio.terra.service.tabulardata.google.bigquery.BigQueryTransactionPdao;
 import bio.terra.stairway.FlightContext;
 import bio.terra.stairway.FlightMap;
 import bio.terra.stairway.Step;
@@ -25,19 +25,19 @@ public class IngestInsertIntoDatasetTableStep implements Step {
       LoggerFactory.getLogger(IngestInsertIntoDatasetTableStep.class);
 
   private final DatasetService datasetService;
-  private final BigQueryPdao bigQueryPdao;
+  private final BigQueryTransactionPdao bigQueryTransactionPdao;
   private final BigQueryDatasetPdao bigQueryDatasetPdao;
   private final AuthenticatedUserRequest userRequest;
   private final boolean autocommit;
 
   public IngestInsertIntoDatasetTableStep(
       DatasetService datasetService,
-      BigQueryPdao bigQueryPdao,
+      BigQueryTransactionPdao bigQueryTransactionPdao,
       BigQueryDatasetPdao bigQueryDatasetPdao,
       AuthenticatedUserRequest userRequest,
       boolean autocommit) {
     this.datasetService = datasetService;
-    this.bigQueryPdao = bigQueryPdao;
+    this.bigQueryTransactionPdao = bigQueryTransactionPdao;
     this.bigQueryDatasetPdao = bigQueryDatasetPdao;
     this.userRequest = userRequest;
     this.autocommit = autocommit;
@@ -100,7 +100,8 @@ public class IngestInsertIntoDatasetTableStep implements Step {
       DatasetTable table = IngestUtils.getDatasetTable(context, dataset);
       UUID transactionId = TransactionUtils.getTransactionId(context);
       try {
-        bigQueryPdao.rollbackDatasetTable(dataset, table.getRawTableName(), transactionId);
+        bigQueryTransactionPdao.rollbackDatasetTable(
+            dataset, table.getRawTableName(), transactionId);
       } catch (InterruptedException e) {
         logger.warn(
             String.format(
@@ -109,7 +110,7 @@ public class IngestInsertIntoDatasetTableStep implements Step {
             e);
       }
       try {
-        bigQueryPdao.rollbackDatasetMetadataTable(dataset, table, transactionId);
+        bigQueryTransactionPdao.rollbackDatasetMetadataTable(dataset, table, transactionId);
       } catch (InterruptedException e) {
         logger.warn(
             String.format(
