+++ src/main/java/bio/terra/service/dataset/flight/transactions/TransactionRollbackDataStep.java
@@ -4,7 +4,7 @@ import bio.terra.service.dataset.Dataset;
 import bio.terra.service.dataset.DatasetService;
 import bio.terra.service.dataset.DatasetTable;
 import bio.terra.service.dataset.flight.ingest.IngestUtils;
-import bio.terra.service.tabulardata.google.bigquery.BigQueryPdao;
+import bio.terra.service.tabulardata.google.bigquery.BigQueryTransactionPdao;
 import bio.terra.stairway.FlightContext;
 import bio.terra.stairway.Step;
 import bio.terra.stairway.StepResult;
@@ -15,13 +15,15 @@ import org.slf4j.LoggerFactory;
 public class TransactionRollbackDataStep implements Step {
   private static final Logger logger = LoggerFactory.getLogger(TransactionRollbackDataStep.class);
   private final DatasetService datasetService;
-  private final BigQueryPdao bigQueryPdao;
+  private final BigQueryTransactionPdao bigQueryTransactionPdao;
   private final UUID transactionId;
 
   public TransactionRollbackDataStep(
-      DatasetService datasetService, BigQueryPdao bigQueryPdao, UUID transactionId) {
+      DatasetService datasetService,
+      BigQueryTransactionPdao bigQueryTransactionPdao,
+      UUID transactionId) {
     this.datasetService = datasetService;
-    this.bigQueryPdao = bigQueryPdao;
+    this.bigQueryTransactionPdao = bigQueryTransactionPdao;
     this.transactionId = transactionId;
   }
 
@@ -29,7 +31,7 @@ public class TransactionRollbackDataStep implements Step {
   public StepResult doStep(FlightContext context) throws InterruptedException {
     Dataset dataset = IngestUtils.getDataset(context, datasetService);
     for (DatasetTable table : dataset.getTables()) {
-      bigQueryPdao.rollbackDatasetTable(dataset, table.getRawTableName(), transactionId);
+      bigQueryTransactionPdao.rollbackDatasetTable(dataset, table.getRawTableName(), transactionId);
     }
     return StepResult.getStepResultSuccess();
   }
