+++ src/main/java/bio/terra/service/dataset/flight/transactions/TransactionRollbackMetadataStep.java
@@ -4,7 +4,7 @@ import bio.terra.service.dataset.Dataset;
 import bio.terra.service.dataset.DatasetService;
 import bio.terra.service.dataset.DatasetTable;
 import bio.terra.service.dataset.flight.ingest.IngestUtils;
-import bio.terra.service.tabulardata.google.bigquery.BigQueryPdao;
+import bio.terra.service.tabulardata.google.bigquery.BigQueryTransactionPdao;
 import bio.terra.stairway.FlightContext;
 import bio.terra.stairway.Step;
 import bio.terra.stairway.StepResult;
@@ -16,13 +16,15 @@ public class TransactionRollbackMetadataStep implements Step {
   private static final Logger logger =
       LoggerFactory.getLogger(TransactionRollbackMetadataStep.class);
   private final DatasetService datasetService;
-  private final BigQueryPdao bigQueryPdao;
+  private final BigQueryTransactionPdao bigQueryTransactionPdao;
   private final UUID transactionId;
 
   public TransactionRollbackMetadataStep(
-      DatasetService datasetService, BigQueryPdao bigQueryPdao, UUID transactionId) {
+      DatasetService datasetService,
+      BigQueryTransactionPdao bigQueryTransactionPdao,
+      UUID transactionId) {
     this.datasetService = datasetService;
-    this.bigQueryPdao = bigQueryPdao;
+    this.bigQueryTransactionPdao = bigQueryTransactionPdao;
     this.transactionId = transactionId;
   }
 
@@ -30,7 +32,7 @@ public class TransactionRollbackMetadataStep implements Step {
   public StepResult doStep(FlightContext context) throws InterruptedException {
     Dataset dataset = IngestUtils.getDataset(context, datasetService);
     for (DatasetTable table : dataset.getTables()) {
-      bigQueryPdao.rollbackDatasetMetadataTable(dataset, table, transactionId);
+      bigQueryTransactionPdao.rollbackDatasetMetadataTable(dataset, table, transactionId);
     }
     return StepResult.getStepResultSuccess();
   }
