+++ src/main/java/bio/terra/service/dataset/flight/transactions/TransactionVerifyStep.java
@@ -5,7 +5,7 @@ import bio.terra.service.dataset.DatasetService;
 import bio.terra.service.dataset.DatasetTable;
 import bio.terra.service.dataset.exception.TransactionCommitException;
 import bio.terra.service.dataset.flight.ingest.IngestUtils;
-import bio.terra.service.tabulardata.google.bigquery.BigQueryPdao;
+import bio.terra.service.tabulardata.google.bigquery.BigQueryTransactionPdao;
 import bio.terra.stairway.FlightContext;
 import bio.terra.stairway.Step;
 import bio.terra.stairway.StepResult;
@@ -19,13 +19,15 @@ import org.slf4j.LoggerFactory;
 public class TransactionVerifyStep implements Step {
   private static final Logger logger = LoggerFactory.getLogger(TransactionVerifyStep.class);
   private final DatasetService datasetService;
-  private final BigQueryPdao bigQueryPdao;
+  private final BigQueryTransactionPdao bigQueryTransactionPdao;
   private final UUID transactionId;
 
   public TransactionVerifyStep(
-      DatasetService datasetService, BigQueryPdao bigQueryPdao, UUID transactionId) {
+      DatasetService datasetService,
+      BigQueryTransactionPdao bigQueryTransactionPdao,
+      UUID transactionId) {
     this.datasetService = datasetService;
-    this.bigQueryPdao = bigQueryPdao;
+    this.bigQueryTransactionPdao = bigQueryTransactionPdao;
     this.transactionId = transactionId;
   }
 
@@ -35,7 +37,8 @@ public class TransactionVerifyStep implements Step {
     List<String> tablesWithConflicts = new ArrayList<>();
     for (DatasetTable datasetTable : dataset.getTables()) {
       if (datasetTable.getPrimaryKey() != null && !datasetTable.getPrimaryKey().isEmpty()) {
-        long conflicts = bigQueryPdao.verifyTransaction(dataset, datasetTable, transactionId);
+        long conflicts =
+            bigQueryTransactionPdao.verifyTransaction(dataset, datasetTable, transactionId);
         if (conflicts > 0) {
           tablesWithConflicts.add(datasetTable.getName());
         }
